chore(provider/kubernetes): de-angularize v2 expected artifact service
