+++ app/scripts/modules/core/src/artifact/artifact.module.ts
@@ -1,7 +1,7 @@
 import { module } from 'angular';
 
 import { EXPECTED_ARTIFACT_SELECTOR_COMPONENT } from './expectedArtifactSelector.component';
-import { EXPECTED_ARTIFACT_SERVICE } from './expectedArtifact.service';
+import { SUMMARIZE_EXPECTED_ARTIFACT_FILTER } from './summarizeExpectedArtifact';
 import { EXPECTED_ARTIFACT_MULTI_SELECTOR_COMPONENT } from './expectedArtifactMultiSelector.component';
 import { IMAGE_SOURCE_SELECTOR_COMPONENT } from './imageSourceSelector.component';
 
@@ -9,6 +9,6 @@ export const ARTIFACT_MODULE = 'spinnaker.core.artifact';
 module(ARTIFACT_MODULE, [
   EXPECTED_ARTIFACT_MULTI_SELECTOR_COMPONENT,
   EXPECTED_ARTIFACT_SELECTOR_COMPONENT,
-  EXPECTED_ARTIFACT_SERVICE,
+  SUMMARIZE_EXPECTED_ARTIFACT_FILTER,
   IMAGE_SOURCE_SELECTOR_COMPONENT,
 ]);
