+++ app/scripts/modules/core/src/artifact/expectedArtifact.service.ts
@@ -1,10 +1,8 @@
-import { copy, module } from 'angular';
-
-import { IArtifact, IExpectedArtifact, IPipeline, IStage } from 'core/domain';
 import { PipelineConfigService } from 'core/pipeline/config/services/PipelineConfigService';
+import { IPipeline, IStage, IExpectedArtifact } from 'core/domain';
 
 export class ExpectedArtifactService {
-  public getExpectedArtifactsAvailableToStage(stage: IStage, pipeline: IPipeline): IExpectedArtifact[] {
+  public static getExpectedArtifactsAvailableToStage(stage: IStage, pipeline: IPipeline): IExpectedArtifact[] {
     let result = pipeline.expectedArtifacts || [];
     PipelineConfigService.getAllUpstreamDependencies(pipeline, stage).forEach(s => {
       const expectedArtifact = (s as any).expectedArtifact;
@@ -20,23 +18,3 @@ export class ExpectedArtifactService {
     return result;
   }
 }
-
-export function summarizeExpectedArtifact() {
-  return function(expected: IExpectedArtifact): string {
-    if (!expected) {
-      return '';
-    }
-
-    const artifact = copy(expected.matchArtifact);
-    return Object.keys(artifact)
-      .filter((k: keyof IArtifact) => artifact[k])
-      .filter(k => k !== 'kind')
-      .map((k: keyof IArtifact) => `${k}: ${artifact[k]}`)
-      .join(', ');
-  };
-}
-
-export const EXPECTED_ARTIFACT_SERVICE = 'spinnaker.core.artifacts.expected.service';
-module(EXPECTED_ARTIFACT_SERVICE, [])
-  .filter('summarizeExpectedArtifact', summarizeExpectedArtifact)
-  .service('expectedArtifactService', ExpectedArtifactService);
