+++ app/scripts/modules/core/src/artifact/expectedArtifactMultiSelector.component.ts
@@ -1,7 +1,5 @@
 import { IComponentOptions, IController, module } from 'angular';
-
 import { IExpectedArtifact } from 'core/domain';
-import { EXPECTED_ARTIFACT_SERVICE, ExpectedArtifactService } from './expectedArtifact.service';
 import { ArtifactIconService } from './ArtifactIconService';
 
 import './artifactSelector.less';
@@ -13,10 +11,6 @@ class ExpectedArtifactMultiSelectorCtrl implements IController {
   public expectedArtifacts: IExpectedArtifact[];
   public helpFieldKey: string;
   public showIcons: boolean;
-
-  constructor(public expectedArtifactService: ExpectedArtifactService) {
-    'ngInject';
-  }
 
   public iconPath(expected: IExpectedArtifact): string {
     const artifact = expected.matchArtifact || expected.defaultArtifact;
@@ -61,7 +55,7 @@ class ExpectedArtifactMultiSelectorComponent implements IComponentOptions {
 }
 
 export const EXPECTED_ARTIFACT_MULTI_SELECTOR_COMPONENT = 'spinnaker.core.artifacts.expected.multiSelector';
-module(EXPECTED_ARTIFACT_MULTI_SELECTOR_COMPONENT, [EXPECTED_ARTIFACT_SERVICE]).component(
+module(EXPECTED_ARTIFACT_MULTI_SELECTOR_COMPONENT, []).component(
   'expectedArtifactMultiSelector',
   new ExpectedArtifactMultiSelectorComponent(),
 );
