+++ app/scripts/modules/core/src/pipeline/config/stages/bakeManifest/bakeManifestConfig.controller.ts
@@ -11,26 +11,28 @@ export class BakeManifestConfigCtrl implements IController {
   public static defaultInputArtifact(): any {
     return {
       id: '',
-      account: ''
-    }
+      account: '',
+    };
   }
 
-  constructor(public $scope: IScope, expectedArtifactService: ExpectedArtifactService) {
+  constructor(public $scope: IScope) {
     'ngInject';
     if (this.$scope.stage.isNew) {
       const defaultSelection = {
         templateRenderer: 'HELM2',
-        expectedArtifacts: [{
-          matchArtifact: {
-            type: 'embedded/base64',
-            kind: 'base64',
-            name: ''
+        expectedArtifacts: [
+          {
+            matchArtifact: {
+              type: 'embedded/base64',
+              kind: 'base64',
+              name: '',
+            },
+            id: UUIDGenerator.generateUuid(),
+            defaultArtifact: {},
+            useDefaultArtifact: false,
           },
-          id: UUIDGenerator.generateUuid(),
-          defaultArtifact: {},
-          useDefaultArtifact: false,
-        }],
-        inputArtifacts: [ BakeManifestConfigCtrl.defaultInputArtifact() ]
+        ],
+        inputArtifacts: [BakeManifestConfigCtrl.defaultInputArtifact()],
       };
 
       Object.assign(this.$scope.stage, defaultSelection);
@@ -40,7 +42,7 @@ export class BakeManifestConfigCtrl implements IController {
       this.artifactAccounts = accounts;
     });
 
-    this.expectedArtifacts = expectedArtifactService.getExpectedArtifactsAvailableToStage(
+    this.expectedArtifacts = ExpectedArtifactService.getExpectedArtifactsAvailableToStage(
       this.$scope.stage,
       this.$scope.$parent.pipeline,
     );
@@ -48,7 +50,7 @@ export class BakeManifestConfigCtrl implements IController {
 
   public hasValueArtifacts(): boolean {
     if (!this.$scope.stage.inputArtifacts) {
-      this.$scope.stage.inputArtifacts = [ BakeManifestConfigCtrl.defaultInputArtifact() ];
+      this.$scope.stage.inputArtifacts = [BakeManifestConfigCtrl.defaultInputArtifact()];
     }
 
     return this.$scope.stage.inputArtifacts.length > 1;
@@ -58,22 +60,22 @@ export class BakeManifestConfigCtrl implements IController {
     // First artifact is special -- the UI depends on it existing. If someone edited the json to remove it,
     // this at least fixes the UI.
     if (!this.$scope.stage.inputArtifacts) {
-      this.$scope.stage.inputArtifacts = [ BakeManifestConfigCtrl.defaultInputArtifact() ];
+      this.$scope.stage.inputArtifacts = [BakeManifestConfigCtrl.defaultInputArtifact()];
     }
 
-    this.$scope.stage.inputArtifacts.push(BakeManifestConfigCtrl.defaultInputArtifact())
+    this.$scope.stage.inputArtifacts.push(BakeManifestConfigCtrl.defaultInputArtifact());
   }
 
   public removeInputArtifact(i: number) {
     if (!this.$scope.stage.inputArtifacts) {
-      this.$scope.stage.inputArtifacts = [ BakeManifestConfigCtrl.defaultInputArtifact() ];
+      this.$scope.stage.inputArtifacts = [BakeManifestConfigCtrl.defaultInputArtifact()];
     }
 
     if (i <= 0) {
       return;
     }
 
-    this.$scope.stage.inputArtifacts.splice(i, 1)
+    this.$scope.stage.inputArtifacts.splice(i, 1);
   }
 
   public outputNameChange() {
